% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aou_execute_cohort}
\alias{aou_execute_cohort}
\title{Execute a cohort from server}
\usage{
aou_execute_cohort(cohortId, baseUrl = "http://atlas-demo.ohdsi.org:80/WebAPI")
}
\arguments{
\item{cohortId}{identifier of cohort on atlas server (in URL of the cohort when defining it)}

\item{baseUrl}{URL of the Atlas server}
}
\value{
cohort id and data.frame with cohort data (person, start and end date)
}
\description{
Execute a cohort from server
}
