% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aou_get_case_report_forms}
\alias{aou_get_case_report_forms}
\title{obtain CRF data dictionary as data.frame using vocabulary tables}
\usage{
aou_get_case_report_forms()
}
\value{
dataframe of CRF dictionary as Answe/Question/Topic/Module
}
\description{
obtain CRF data dictionary as data.frame using vocabulary tables
}
