% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aou_tbl_fields}
\alias{aou_tbl_fields}
\title{List columns of a table}
\usage{
aou_tbl_fields(table)
}
\arguments{
\item{table}{name of table (e.g., person)}
}
\value{
data frame with fields and type and description
}
\description{
List columns of a table
}
