% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aou_write}
\alias{aou_write}
\title{Write data into bucket}
\usage{
aou_write(data, destination_filename)
}
\arguments{
\item{data}{data frame with the data}

\item{destination_filename}{name of the file to store the data}
}
\value{

}
\description{
Write data into bucket
}
